подправил стиль кода

+++ src/main/java/com/logstream/korobkin/library/controllers/BooksController.java
@@ -1,10 +1,7 @@
 package com.logstream.korobkin.library.controllers;
 
-
 import com.logstream.korobkin.library.models.Book;
-
-
-import com.logstream.korobkin.library.services.BookService;
+import com.logstream.korobkin.library.services.BooksService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,29 +17,28 @@ import java.util.List;
 @RequestMapping()
 public class BooksController {
 
-    private final BookService bookService;
+    private final BooksService bookService;
 
     @Autowired
-    public BooksController(BookService bookService) {
+    public BooksController(BooksService bookService) {
         this.bookService = bookService;
     }
 
-
     @GetMapping("/all")
     @Transactional(readOnly = true)
-    public List<Book> showBooks(Model model){
-        return  bookService.findAll();
+    public List<Book> showBooks(Model model) {
+        return bookService.findAll();
     }
 
     @GetMapping("/isbn/{isbn}")
     @Transactional(readOnly = true)
-    public Book searchBook(@PathVariable("isbn") String isbn,Model model){
+    public Book searchBook(@PathVariable("isbn") String isbn, Model model) {
         return bookService.findOne(isbn);
     }
 
     @PostMapping("/{isbn}")
-    public ResponseEntity<HttpStatus> newBook(@RequestBody @Valid Book book, BindingResult bindingResult){
-        if(bindingResult.hasErrors()){
+    public ResponseEntity<HttpStatus> newBook(@RequestBody @Valid Book book, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(HttpStatus.BAD_REQUEST);
         }
         bookService.save(book);
@@ -50,35 +46,17 @@ public class BooksController {
     }
 
     @PostMapping("/isbn/{isbn}")
-    public List<Book>updateBook(@PathVariable("isbn") String isbn, @ModelAttribute("book") Book book){
+    public List<Book> updateBook(@PathVariable("isbn") String isbn, @ModelAttribute("book") Book book) {
         bookService.update(isbn, book);
         return bookService.findAll();
     }
 
     @GetMapping("/delete/{isbn}")
-    public List<Book> deleteBook(@RequestBody @Valid Book book){
-        bookService.delete(book);
-        return  bookService.findAll();
+    public List<Book> deleteBook(@PathVariable("isbn") String isbn) {
+        Book book = bookService.findOne(isbn);
+        if (book != null) {
+            bookService.delete(book);
+        }
+        return bookService.findAll();
     }
-
-
-//    @Transactional
-//   public Optional<Book> findBookByIsbn(String isbn) {
-//        return booksRepository.findByIsbn(isbn);
-//   }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
